+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed:
   - Usage of `Api:api` with not fully qualified class name
   - Usage of `Api::HTTP_RESPONSE_*` constants
+  - Usage of response format other than `json`
 
 ## 2.2.0 / Unreleased
 ### Added: